requirements.txt
aiogram==2.25.1
tensorflow==2.13.0
keras==2.13.1
pillow==10.1.0
numpy==1.24.3
python-dotenv==1.0.0